fix: GitHubUtils
