+++ plugins/github/server/github.ts
@@ -36,27 +36,6 @@ export class GitHub {
   }
 
   /**
-   * @param error
-   * @returns URL to be redirected to upon authorization error from GitHub
-   */
-  public static errorUrl(error: string) {
-    return `${this.url}?error=${error}`;
-  }
-
-  /**
-   * @returns Callback URL configured for GitHub, to which users will be redirected upon authorization
-   */
-  public static callbackUrl({
-    baseUrl = `${env.URL}/api/github.callback`,
-    params,
-  }: {
-    baseUrl: string;
-    params?: string;
-  }) {
-    return `${baseUrl}/api/github.callback?${params}`;
-  }
-
-  /**
    * @param installationId Identifies a GitHub App installation
    * @returns {object} An object containing details about the GitHub App installation,
    * e.g, installation target, account which installed the app etc.
